safari-600.7-branch/LayoutTests/fast/xpath/nsresolver-bad-object-expected.txt

CONSOLE MESSAGE: XPathNSResolver does not have a lookupNamespaceURI method.
SUCCESS
